[refactor] MapScreen 내에 Component 분리
